examples/python/cuDF/requirements.txt
cachetools==5.3.3
cuda-python==12.4.0
cudf-cu12==24.4.0
cupy-cuda12x==13.0.0
fastrlock==0.8.2
fsspec==2024.3.1
llvmlite==0.42.0
markdown-it-py==3.0.0
mdurl==0.1.2
numba==0.59.1
numpy==1.26.4
nvtx==0.2.10
packaging==24.0
pandas==2.2.1
protobuf==4.25.3
pyarrow==14.0.2
Pygments==2.17.2
pynvjitlink-cu12==0.1.14
python-dateutil==2.9.0.post0
pytz==2024.1
rich==13.7.1
rmm-cu12==24.4.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
